requirements.txt
sf-hamilton>=1.52.0
pandas>=2.0.0
pyarrow>=14.0.0
matplotlib>=3.8.0
requests>=2.31.0
sseclient-py>=1.8.0
nbformat>=5.9.0
openpyxl>=3.1.0
scikit-learn>=1.0.0
seaborn>=0.12.0
gseapy>=1.0.0
scipy>=1.11.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.1.0
reportlab>=4.0.0
